Still messing with it. There are improvements

+++ No-G Unity/Assets/Scripts/PlayerHealth.cs
@@ -6,7 +6,7 @@ using Photon.Pun;
 
 public class PlayerHealth : MonoBehaviourPun, IPunObservable
 {
-    public float score = 0;
+    public int score = 0;
 
     public bool playerJustJoined = true;
     public bool isDead = false;
